[IMP] point_of_sale: new class for payment validation in PoS.
We need now to adapt each sub module and loca.
