+++ addons/point_of_sale/static/src/app/screens/feedback_screen/feedback_screen.js
@@ -17,24 +17,14 @@ export class FeedbackScreen extends Component {
         this.pos = usePos();
         this.containerRef = useRef("feedback-screen");
         this.amountRef = useRef("amount");
-        onMounted(async () => {
+        onMounted(() => {
             this.scaleText();
-            if (this.pos.isFastPaymentRunning) {
-                await this.fastValidate();
-            }
         });
         if (!this.pos.isFastPaymentRunning) {
             this.timeout = setTimeout(() => {
                 this.goToNextScreen();
             }, 2000);
         }
-        this.virtualPaymentScreen = useRef("virtualPaymentScreen");
-    }
-
-    async fastValidate() {
-        const paymentMethod = this.pos.models["pos.payment.method"].get(this.props.paymentMethodId);
-        await this.pos.validateOrderFast(this.virtualPaymentScreen.el, paymentMethod);
-        this.goToNextScreen();
     }
 
     scaleText() {
